add: refactor compose_variable thus finishing negation!
