+++ src/main.rs
@@ -114,7 +114,7 @@ fn english_to_agda(knowledge: Vec<String>, conclusions: Vec<String>) -> AgdaFile
 
 #[tokio::main]
 async fn main() {
-    let config = Config::from_args("john is a big man");
+    let config = Config::from_args("every john is not jack");
     let knowledge = config.knowledge;
     let conclusions = config.conclusions;
